Выполняет задание 5.15 Модуляция.
