+++ js/constants.js
@@ -1,0 +1,71 @@
+const NAMES = [
+  'Константин',
+  'Олег',
+  'Михаил',
+  'Александра',
+  'Соня',
+  'Егор',
+  'Дмитрий',
+  'Яна',
+  'Ксения',
+  'Юлия',
+  'Злата',
+  'Сергей',
+  'Виталий',
+  'Ольга',
+  'Роман',
+  'Олег',
+  'Глеб',
+  'Марина',
+  'Ирина',
+  'Ева',
+  'Алиса',
+  'Анелина',
+  'Алина',
+  'Арина',
+  'Инга',
+];
+
+const SURNAMES = [
+  'Смирнов',
+  'Новиков',
+  'Петров',
+  'Максимова',
+  'Гаврилова',
+  'Ершов',
+  'Исаев',
+  'Осипова',
+  'Третьякова',
+  'Евдокимова',
+  'Аксёнов',
+  'Николаев',
+  'Моисеев',
+  'Лапина',
+  'Дьячков',
+  'Фролов',
+  'Одинцов',
+  'Трофимова',
+  'Нестерова',
+  'Филатова',
+  'Авдеева',
+  'Антонова',
+  'Назарова',
+  'Белова',
+  'Селезнёва',
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const MAX_COMMENTS_COUNT = 6;
+const MIN_COMMENTS_COUNT = 3;
+const MAX_LIKES_COUNT = 25;
+const PHOTO_COUNT = 25;
+
+export {NAMES, SURNAMES, MESSAGE, MAX_COMMENTS_COUNT, MIN_COMMENTS_COUNT, MAX_LIKES_COUNT, PHOTO_COUNT};
